refactor(json): unify JSON schema across benchmark modes
