+++ MANUAL.md
@@ -473,7 +473,7 @@ For noisy systems, prioritize median and P95/P99 rather than single fastest/slow
   "main_memory": { ... },
   "cache": { ... },
   "timestamp": "2026-03-09T14:57:56Z",
-  "version": "0.53.1"
+  "version": "0.53.7"
 }
 ```
 
@@ -485,7 +485,7 @@ For noisy systems, prioritize median and P95/P99 rather than single fastest/slow
   "execution_time_sec": 705.6,
   "patterns": { ... },
   "timestamp": "2026-03-09T15:10:01Z",
-  "version": "0.53.1"
+  "version": "0.53.7"
 }
 ```
 
@@ -510,16 +510,18 @@ When `-latency-stride-bytes` is explicitly set, latency sections also include `c
       "max": 27.73
     }
   },
-  "samples_ns": [26.86, 26.77, 26.69],
-  "samples_statistics": {
-    "average": 26.75,
-    "median": 26.72,
-    "p90": 27.30,
-    "p95": 27.51,
-    "p99": 27.68,
-    "stddev": 0.47,
-    "min": 26.24,
-    "max": 27.73
+  "samples_ns": {
+    "values": [26.86, 26.77, 26.69],
+    "statistics": {
+      "average": 26.75,
+      "median": 26.72,
+      "p90": 27.30,
+      "p95": 27.51,
+      "p99": 27.68,
+      "stddev": 0.47,
+      "min": 26.24,
+      "max": 27.73
+    }
   },
   "auto_tlb_breakdown": {
     "tlb_hit_ns": {
@@ -574,18 +576,19 @@ When `-latency-stride-bytes` is explicitly set, latency sections also include `c
 ### TLB analysis JSON (analyze mode)
 
 When run with `-analyze-tlb -output tlb_analysis.json`, the payload includes a dedicated `tlb_analysis` block.
-Example below is modeled from `results/macminim4_analyte-tbl_stride_136B.json`:
+Example below is modeled from `results/0.53.7/MacMiniM4_analyzetlb.json`:
 
 ```json
 {
   "configuration": {
     "mode": "analyze_tlb",
-    "latency_stride_bytes": 136,
+    "latency_stride_bytes": 64,
     "buffer_size_mb": 1024
   },
   "tlb_analysis": {
     "sweep": [
       {
+        "locality_bytes": 16384,
         "locality_kb": 16,
         "loop_latencies_ns": [25.957278, 25.965990, 25.916902],
         "p50_latency_ns": 25.982678
@@ -635,7 +638,7 @@ python3 -m json.tool results.json
 jq '.main_memory.bandwidth.read_gb_s.statistics.median' results.json
 
 # Main memory latency P95 from sample distribution
-jq '.main_memory.latency.samples_statistics.p95' results.json
+jq '.main_memory.latency.samples_ns.statistics.p95' results.json
 
 # Auto TLB miss latency median (when auto comparison is active)
 jq '.main_memory.latency.auto_tlb_breakdown.tlb_miss_ns.statistics.median' results.json
@@ -661,7 +664,7 @@ What it does:
 - Sweeps multiple custom cache sizes
 - Sweeps multiple `-latency-tlb-locality-kb` values
 - Writes per-run JSON files under `script-examples/tmp/`
-- Extracts `.cache.custom.latency.samples_statistics` into `script-examples/final_output.txt`
+- Extracts `.cache.custom.latency.samples_ns.statistics` into `script-examples/final_output.txt`
 - Clears `tmp` after extraction
 
 Important: the script currently invokes `memory_benchmark` from `PATH`. If you only built locally as `./memory_benchmark`, either install it to `PATH` or update `BENCHMARK_CMD` in the script.
@@ -771,8 +774,10 @@ Make sure you are passing `script-examples/final_output.txt` generated by the la
 
 Repository sample result files:
 
-- `results/macminim4_benchmark_count10.json`
-- `results/macminim4_patterns_count10.json`
+- `results/0.53.7/MacMiniM4_benchmark.json`
+- `results/0.53.7/MacMiniM4_patterns.json`
+- `results/0.53.7/MacMiniM4_analyzetlb.json`
+- `results/0.53.7/MacMiniM4_core2core.json`
 
 Command help:
 
@@ -782,4 +787,4 @@ Command help:
 
 ---
 
-**Last Updated**: 2026-03-12
+**Last Updated**: 2026-03-13
