+++ script-examples/latency_test_script.sh
@@ -116,7 +116,7 @@ done
 
 echo ""
 echo "=========================================="
-echo "Extracting samples_statistics from output files..."
+echo "Extracting samples_ns.statistics from output files..."
 echo ""
 
 # Output file for aggregated statistics (in script directory)
@@ -125,18 +125,18 @@ final_output="${SCRIPT_DIR}/final_output.txt"
 # Clear/create the final output file
 > "${final_output}"
 
-# Function to extract samples_statistics using jq
+# Function to extract samples_ns.statistics using jq
 extract_with_jq() {
     local json_file=$1
     local cache_size=$2
     local tlb_kb=$3
     echo "TLB Locality: ${tlb_kb} KB, Cache Size: ${cache_size} KB" >> "${final_output}"
     echo "----------------------------------------" >> "${final_output}"
-    jq '.cache.custom.latency.samples_statistics' "${json_file}" >> "${final_output}"
+    jq '.cache.custom.latency.samples_ns.statistics' "${json_file}" >> "${final_output}"
     echo "" >> "${final_output}"
 }
 
-# Function to extract samples_statistics using Python
+# Function to extract samples_ns.statistics using Python
 extract_with_python() {
     local json_file=$1
     local cache_size=$2
@@ -149,7 +149,7 @@ import sys
 try:
     with open('${json_file}', 'r') as f:
         data = json.load(f)
-        stats = data['cache']['custom']['latency']['samples_statistics']
+        stats = data['cache']['custom']['latency']['samples_ns']['statistics']
         print(json.dumps(stats, indent=2))
 except Exception as e:
     print(f'Error: {e}', file=sys.stderr)
@@ -160,17 +160,17 @@ EOF
 
 # Check if jq is available, otherwise use Python
 if command -v jq &> /dev/null; then
-    echo "Using jq to extract samples_statistics..."
+    echo "Using jq to extract samples_ns.statistics..."
     USE_JQ=true
 elif command -v python3 &> /dev/null; then
-    echo "Using Python to extract samples_statistics..."
+    echo "Using Python to extract samples_ns.statistics..."
     USE_JQ=false
 else
-    echo "Error: Neither jq nor python3 is available. Cannot extract samples_statistics."
+    echo "Error: Neither jq nor python3 is available. Cannot extract samples_ns.statistics."
     exit 1
 fi
 
-# Extract samples_statistics from each output file
+# Extract samples_ns.statistics from each output file
 for tlb_kb in "${tlb_locality_sizes_kb[@]}"; do
     for cache_size in "${cache_sizes[@]}"; do
         output_file="${TMP_DIR}/output_tlb_${tlb_kb}_cache_${cache_size}.json"
@@ -188,7 +188,7 @@ for tlb_kb in "${tlb_locality_sizes_kb[@]}"; do
 done
 
 echo "=========================================="
-echo "All samples_statistics extracted to ${final_output}"
+echo "All samples_ns.statistics extracted to ${final_output}"
 
 # Clear tmp folder after final_output.txt is created
 echo ""
